regenerate fields.yml from scratch
this one is interesting. only CI could reproduce this diff, when running
Make, it generated this diff. But locally, couldn't make this happen.
Even booting an ubuntu:20.04 docker and cloning the repo from scratch
did not reproduce this diff. But running
find package/endpoint -name fields.yml -delete
to fully delete these, and then running make, did it recreate the exact
diff that CI saw. I wonder if these should be deleted and rebuilt more
often, or as part of the Makefile